fix(ai-world): widen money to uint64/BIGINT to remove overflow ceiling (2.11E2 P3)
Money was the last outstanding non-blocking P3: uint32/INT UNSIGNED
gave it a wraparound ceiling that a high enough
AIWorld.WorkMoneyReward or enough reward cycles could reach. Widened
to uint64 in AgentEconomyState and BIGINT UNSIGNED in ai_agents (a
plain MODIFY COLUMN - every existing value fits unchanged). Food/
Resource stay uint32 for now; nothing mutates them yet, so there is
nothing to overflow - widen them the same way once a milestone
actually starts writing to them.

The other previously-flagged P3 (async economy writes needing a more
robust ordering model once more mutation types write more often) is
already covered by the economy_version monotonic guard added in the
prior commit - Version is a single counter across the whole
AgentEconomyState row, not specific to the WORK reward, so it already
generalizes to any number of mutation types on the single world
thread; no further change needed for that here.

+++ src/server/game/AIWorld/Agent/AgentEconomyState.h
@@ -32,7 +32,13 @@
 // schema migration, but nothing writes to them yet.
 struct AgentEconomyState
 {
-    uint32 Money = 0;
+    // Milestone 2.11E2 P3 fix: uint64/BIGINT UNSIGNED, not uint32/INT
+    // UNSIGNED - the only field anything currently mutates, so it is the
+    // only one actually at risk of wrapping back to a small value after
+    // enough WORK rewards. Food/Resource stay uint32 for now: nothing
+    // writes to them yet, so there is nothing to overflow - widen them the
+    // same way whenever a milestone actually starts mutating them.
+    uint64 Money = 0;
     uint32 Food = 0;
     uint32 Resource = 0;
